@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
-  appVersion: '1.0.5',
+  appVersion: '2.0.0',
   apiBaseUrl: 'https://api.diecasttracker.com'
 };
